pyproject.toml
[tool.poetry]
name = "reez"
version = "0.1.0"
description = "X-Ray Registration System"
authors = ["resakse <resakse@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
django-debug-toolbar = "^4.2.0"
django-crispy-forms = "^2.0"
django-filter = "^23.3"
django-extensions = "^3.2.3"
django-htmx = "^1.17.0"
crispy-bootstrap5 = "^0.7"
django-auto-prefetch = "^1.8.0"
django-ordered-model = "^3.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
